refactor(runtime): extract workflow / knowledge / schedule / cron tool modules
Carve out four more leaf domains from tool_runner/mod.rs:

- `workflow` — run / list / status / start / cancel
- `knowledge` — add_entity / add_relation / query, plus the
  parse_entity_type / parse_relation_type helpers they share.
- `schedule` — natural-language schedule wrappers create / list / delete,
  with the parse_schedule_to_cron and parse_time_to_hour helpers kept
  module-private (they exist only to serve the schedule_* surface).
- `cron` — low-level cron_create / list / cancel that go straight to
  KernelHandle::cron_*. Includes the ownership check that keeps
  cron_cancel from deleting another agent's jobs.

Each tool keeps its original signature and body verbatim; only the
`fn` -> `pub(super) fn` visibility tweak was needed so the parent
dispatch sites can keep their existing call shape.

mod.rs: 13,123 -> 12,510 LOC (-613). cargo check + cargo clippy --lib
both clean.

Refs #3710
