+++ crates/librefang-runtime/src/tool_runner/mod.rs
@@ -15,21 +15,30 @@ use std::sync::Arc;
 use tracing::{debug, warn};
 
 mod agent;
+mod cron;
 mod event;
 mod goal;
+mod knowledge;
 mod memory;
 mod notify;
+mod schedule;
 mod shell_safety;
 mod system;
 mod taint;
 mod task;
 mod wiki;
+mod workflow;
 
 use self::agent::tool_agent_find;
+use self::cron::{tool_cron_cancel, tool_cron_create, tool_cron_list};
 use self::event::tool_event_publish;
 use self::goal::tool_goal_update;
+use self::knowledge::{
+    tool_knowledge_add_entity, tool_knowledge_add_relation, tool_knowledge_query,
+};
 use self::memory::{tool_memory_list, tool_memory_recall, tool_memory_store};
 use self::notify::tool_notify_owner;
+use self::schedule::{tool_schedule_create, tool_schedule_delete, tool_schedule_list};
 use self::shell_safety::{classify_shell_exec_ro_safety, RoSafety};
 use self::system::{tool_location_get, tool_system_time};
 use self::taint::{
@@ -40,6 +49,10 @@ use self::task::{
     tool_task_claim, tool_task_complete, tool_task_list, tool_task_post, tool_task_status,
 };
 use self::wiki::{tool_wiki_get, tool_wiki_search, tool_wiki_write};
+use self::workflow::{
+    tool_workflow_cancel, tool_workflow_list, tool_workflow_run, tool_workflow_start,
+    tool_workflow_status,
+};
 
 /// Maximum inter-agent call depth to prevent infinite recursion (A->B->C->...).
 #[allow(dead_code)]
@@ -3894,619 +3907,6 @@ fn tool_meta_search(
             lines.join("\n")
         ),
     )
-}
-
-// ---------------------------------------------------------------------------
-// Workflow execution tool
-// ---------------------------------------------------------------------------
-
-async fn tool_workflow_run(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let workflow_id = input["workflow_id"]
-        .as_str()
-        .ok_or("Missing 'workflow_id' parameter")?;
-
-    // Serialize optional input object to a JSON string for the workflow engine.
-    let input_str = match input.get("input") {
-        Some(v) if v.is_object() => serde_json::to_string(v)
-            .map_err(|e| format!("Failed to serialize workflow input: {e}"))?,
-        Some(v) if v.is_null() => String::new(),
-        Some(_) => return Err("'input' must be a JSON object or null".to_string()),
-        None => String::new(),
-    };
-
-    let kh = require_kernel(kernel)?;
-    let (run_id, output) = kh
-        .run_workflow(workflow_id, &input_str)
-        .await
-        .map_err(|e| e.to_string())?;
-
-    Ok(serde_json::json!({
-        "run_id": run_id,
-        "output": output,
-    })
-    .to_string())
-}
-
-// ---------------------------------------------------------------------------
-// workflow_list — enumerate registered workflow definitions
-// ---------------------------------------------------------------------------
-
-async fn tool_workflow_list(kernel: Option<&Arc<dyn KernelHandle>>) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let mut summaries = kh.list_workflows().await;
-    // Sort by name for deterministic LLM prompt output (#3298).
-    summaries.sort_by(|a, b| a.name.cmp(&b.name));
-    let json_array: Vec<serde_json::Value> = summaries
-        .into_iter()
-        .map(|w| {
-            serde_json::json!({
-                "id": w.id,
-                "name": w.name,
-                "description": w.description,
-                "step_count": w.step_count,
-            })
-        })
-        .collect();
-    serde_json::to_string(&json_array)
-        .map_err(|e| format!("Failed to serialize workflow list: {e}"))
-}
-
-// ---------------------------------------------------------------------------
-// workflow_status — get the status of a workflow run
-// ---------------------------------------------------------------------------
-
-async fn tool_workflow_status(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let run_id = input["run_id"]
-        .as_str()
-        .ok_or("Missing 'run_id' parameter")?;
-
-    // Validate UUID format before calling kernel — returns a clear error
-    // rather than silently returning not-found for a malformed id.
-    uuid::Uuid::parse_str(run_id)
-        .map_err(|_| format!("Invalid run_id — must be a UUID: {run_id}"))?;
-
-    let kh = require_kernel(kernel)?;
-    let summary = kh
-        .get_workflow_run(run_id)
-        .await
-        .ok_or_else(|| format!("workflow run not found: {run_id}"))?;
-
-    serde_json::to_string(&serde_json::json!({
-        "run_id": summary.run_id,
-        "workflow_id": summary.workflow_id,
-        "workflow_name": summary.workflow_name,
-        "state": summary.state,
-        "started_at": summary.started_at,
-        "completed_at": summary.completed_at,
-        "output": summary.output,
-        "error": summary.error,
-        "step_count": summary.step_count,
-        "last_step_name": summary.last_step_name,
-    }))
-    .map_err(|e| format!("Failed to serialize workflow status: {e}"))
-}
-
-// ---------------------------------------------------------------------------
-// workflow_start — fire-and-forget async workflow launch
-// ---------------------------------------------------------------------------
-
-async fn tool_workflow_start(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let workflow_id = input["workflow_id"]
-        .as_str()
-        .ok_or("Missing 'workflow_id' parameter")?;
-
-    // Serialize optional input object to a JSON string for the workflow engine.
-    let input_str = match input.get("input") {
-        Some(v) if v.is_object() => serde_json::to_string(v)
-            .map_err(|e| format!("Failed to serialize workflow input: {e}"))?,
-        Some(v) if v.is_null() => String::new(),
-        Some(_) => return Err("'input' must be a JSON object or null".to_string()),
-        None => String::new(),
-    };
-
-    let kh = require_kernel(kernel)?;
-    let run_id = kh
-        .start_workflow_async(workflow_id, &input_str)
-        .await
-        .map_err(|e| e.to_string())?;
-
-    Ok(serde_json::json!({ "run_id": run_id }).to_string())
-}
-
-// ---------------------------------------------------------------------------
-// workflow_cancel — cancel a running or paused workflow run
-// ---------------------------------------------------------------------------
-
-async fn tool_workflow_cancel(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let run_id = input["run_id"]
-        .as_str()
-        .ok_or("Missing 'run_id' parameter")?;
-
-    // Validate UUID format before calling kernel.
-    uuid::Uuid::parse_str(run_id)
-        .map_err(|_| format!("Invalid run_id — must be a UUID: {run_id}"))?;
-
-    let kh = require_kernel(kernel)?;
-    kh.cancel_workflow_run(run_id)
-        .await
-        .map_err(|e| e.to_string())?;
-
-    Ok(serde_json::json!({
-        "run_id": run_id,
-        "state": "cancelled",
-    })
-    .to_string())
-}
-
-// ---------------------------------------------------------------------------
-// Knowledge graph tools
-// ---------------------------------------------------------------------------
-
-fn parse_entity_type(s: &str) -> librefang_types::memory::EntityType {
-    use librefang_types::memory::EntityType;
-    match s.to_lowercase().as_str() {
-        "person" => EntityType::Person,
-        "organization" | "org" => EntityType::Organization,
-        "project" => EntityType::Project,
-        "concept" => EntityType::Concept,
-        "event" => EntityType::Event,
-        "location" => EntityType::Location,
-        "document" | "doc" => EntityType::Document,
-        "tool" => EntityType::Tool,
-        other => EntityType::Custom(other.to_string()),
-    }
-}
-
-fn parse_relation_type(s: &str) -> librefang_types::memory::RelationType {
-    use librefang_types::memory::RelationType;
-    match s.to_lowercase().as_str() {
-        "works_at" | "worksat" => RelationType::WorksAt,
-        "knows_about" | "knowsabout" | "knows" => RelationType::KnowsAbout,
-        "related_to" | "relatedto" | "related" => RelationType::RelatedTo,
-        "depends_on" | "dependson" | "depends" => RelationType::DependsOn,
-        "owned_by" | "ownedby" => RelationType::OwnedBy,
-        "created_by" | "createdby" => RelationType::CreatedBy,
-        "located_in" | "locatedin" => RelationType::LocatedIn,
-        "part_of" | "partof" => RelationType::PartOf,
-        "uses" => RelationType::Uses,
-        "produces" => RelationType::Produces,
-        other => RelationType::Custom(other.to_string()),
-    }
-}
-
-async fn tool_knowledge_add_entity(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let name = input["name"].as_str().ok_or("Missing 'name' parameter")?;
-    let entity_type_str = input["entity_type"]
-        .as_str()
-        .ok_or("Missing 'entity_type' parameter")?;
-    let properties = input
-        .get("properties")
-        .and_then(|v| v.as_object())
-        .map(|m| m.iter().map(|(k, v)| (k.clone(), v.clone())).collect())
-        .unwrap_or_default();
-
-    let entity = librefang_types::memory::Entity {
-        id: String::new(), // kernel/store assigns a real ID
-        entity_type: parse_entity_type(entity_type_str),
-        name: name.to_string(),
-        properties,
-        created_at: chrono::Utc::now(),
-        updated_at: chrono::Utc::now(),
-    };
-
-    let id = kh
-        .knowledge_add_entity(&entity)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(format!("Entity '{name}' added with ID: {id}"))
-}
-
-async fn tool_knowledge_add_relation(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let source = input["source"]
-        .as_str()
-        .ok_or("Missing 'source' parameter")?;
-    let relation_str = input["relation"]
-        .as_str()
-        .ok_or("Missing 'relation' parameter")?;
-    let target = input["target"]
-        .as_str()
-        .ok_or("Missing 'target' parameter")?;
-    let confidence = input["confidence"].as_f64().unwrap_or(1.0) as f32;
-    let properties = input
-        .get("properties")
-        .and_then(|v| v.as_object())
-        .map(|m| m.iter().map(|(k, v)| (k.clone(), v.clone())).collect())
-        .unwrap_or_default();
-
-    let relation = librefang_types::memory::Relation {
-        source: source.to_string(),
-        relation: parse_relation_type(relation_str),
-        target: target.to_string(),
-        properties,
-        confidence,
-        created_at: chrono::Utc::now(),
-    };
-
-    let id = kh
-        .knowledge_add_relation(&relation)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(format!(
-        "Relation '{source}' --[{relation_str}]--> '{target}' added with ID: {id}"
-    ))
-}
-
-async fn tool_knowledge_query(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let source = input["source"].as_str().map(|s| s.to_string());
-    let target = input["target"].as_str().map(|s| s.to_string());
-    let relation = input["relation"].as_str().map(parse_relation_type);
-    // Cap depth to prevent LLM-triggered DoS via exponential graph
-    // traversal. Knowledge graphs rarely benefit from depth > 5 and
-    // the backend traversal is O(branching_factor^depth).
-    const MAX_KNOWLEDGE_DEPTH: u64 = 10;
-    let max_depth = input["max_depth"]
-        .as_u64()
-        .unwrap_or(1)
-        .min(MAX_KNOWLEDGE_DEPTH) as u32;
-
-    let pattern = librefang_types::memory::GraphPattern {
-        source,
-        relation,
-        target,
-        max_depth,
-    };
-
-    let matches = kh
-        .knowledge_query(pattern)
-        .await
-        .map_err(|e| e.to_string())?;
-    if matches.is_empty() {
-        return Ok("No matching knowledge graph entries found.".to_string());
-    }
-
-    let mut output = format!("Found {} match(es):\n", matches.len());
-    for m in &matches {
-        output.push_str(&format!(
-            "\n  {} ({:?}) --[{:?} ({:.0}%)]--> {} ({:?})",
-            m.source.name,
-            m.source.entity_type,
-            m.relation.relation,
-            m.relation.confidence * 100.0,
-            m.target.name,
-            m.target.entity_type,
-        ));
-    }
-    Ok(output)
-}
-
-// ---------------------------------------------------------------------------
-// Scheduling tools
-// ---------------------------------------------------------------------------
-
-/// Parse a natural language schedule into a cron expression.
-fn parse_schedule_to_cron(input: &str) -> Result<String, String> {
-    let input = input.trim().to_lowercase();
-
-    // If it already looks like a cron expression (5 space-separated fields), pass through
-    let parts: Vec<&str> = input.split_whitespace().collect();
-    if parts.len() == 5
-        && parts
-            .iter()
-            .all(|p| p.chars().all(|c| c.is_ascii_digit() || "*/,-".contains(c)))
-    {
-        return Ok(input);
-    }
-
-    // Natural language patterns
-    if let Some(rest) = input.strip_prefix("every ") {
-        if rest == "minute" || rest == "1 minute" {
-            return Ok("* * * * *".to_string());
-        }
-        if let Some(mins) = rest.strip_suffix(" minutes") {
-            let n: u32 = mins
-                .trim()
-                .parse()
-                .map_err(|_| format!("Invalid number in '{input}'"))?;
-            if n == 0 || n > 59 {
-                return Err(format!("Minutes must be 1-59, got {n}"));
-            }
-            return Ok(format!("*/{n} * * * *"));
-        }
-        if rest == "hour" || rest == "1 hour" {
-            return Ok("0 * * * *".to_string());
-        }
-        if let Some(hrs) = rest.strip_suffix(" hours") {
-            let n: u32 = hrs
-                .trim()
-                .parse()
-                .map_err(|_| format!("Invalid number in '{input}'"))?;
-            if n == 0 || n > 23 {
-                return Err(format!("Hours must be 1-23, got {n}"));
-            }
-            return Ok(format!("0 */{n} * * *"));
-        }
-        if rest == "day" || rest == "1 day" {
-            return Ok("0 0 * * *".to_string());
-        }
-        if rest == "week" || rest == "1 week" {
-            return Ok("0 0 * * 0".to_string());
-        }
-    }
-
-    // "daily at Xam/pm"
-    if let Some(time_str) = input.strip_prefix("daily at ") {
-        let hour = parse_time_to_hour(time_str)?;
-        return Ok(format!("0 {hour} * * *"));
-    }
-
-    // "weekdays at Xam/pm"
-    if let Some(time_str) = input.strip_prefix("weekdays at ") {
-        let hour = parse_time_to_hour(time_str)?;
-        return Ok(format!("0 {hour} * * 1-5"));
-    }
-
-    // "weekends at Xam/pm"
-    if let Some(time_str) = input.strip_prefix("weekends at ") {
-        let hour = parse_time_to_hour(time_str)?;
-        return Ok(format!("0 {hour} * * 0,6"));
-    }
-
-    // "hourly" / "daily" / "weekly" / "monthly"
-    match input.as_str() {
-        "hourly" => return Ok("0 * * * *".to_string()),
-        "daily" => return Ok("0 0 * * *".to_string()),
-        "weekly" => return Ok("0 0 * * 0".to_string()),
-        "monthly" => return Ok("0 0 1 * *".to_string()),
-        _ => {}
-    }
-
-    Err(format!(
-        "Could not parse schedule '{input}'. Try: 'every 5 minutes', 'daily at 9am', 'weekdays at 6pm', or a cron expression like '0 */5 * * *'"
-    ))
-}
-
-/// Parse a time string like "9am", "6pm", "14:00", "9:30am" into an hour (0-23).
-fn parse_time_to_hour(s: &str) -> Result<u32, String> {
-    let s = s.trim().to_lowercase();
-
-    // Handle "9am", "6pm", "12pm", "12am"
-    if let Some(h) = s.strip_suffix("am") {
-        let hour: u32 = h.trim().parse().map_err(|_| format!("Invalid time: {s}"))?;
-        return match hour {
-            12 => Ok(0),
-            1..=11 => Ok(hour),
-            _ => Err(format!("Invalid hour: {hour}")),
-        };
-    }
-    if let Some(h) = s.strip_suffix("pm") {
-        let hour: u32 = h.trim().parse().map_err(|_| format!("Invalid time: {s}"))?;
-        return match hour {
-            12 => Ok(12),
-            1..=11 => Ok(hour + 12),
-            _ => Err(format!("Invalid hour: {hour}")),
-        };
-    }
-
-    // Handle "14:00" or "9:30"
-    if let Some((h, _m)) = s.split_once(':') {
-        let hour: u32 = h.trim().parse().map_err(|_| format!("Invalid time: {s}"))?;
-        if hour > 23 {
-            return Err(format!("Hour must be 0-23, got {hour}"));
-        }
-        return Ok(hour);
-    }
-
-    // Plain number
-    let hour: u32 = s.parse().map_err(|_| format!("Invalid time: {s}"))?;
-    if hour > 23 {
-        return Err(format!("Hour must be 0-23, got {hour}"));
-    }
-    Ok(hour)
-}
-
-// schedule_* tools — high-level wrappers around the CronScheduler engine.
-// These accept natural language schedules ("daily at 9am") and delegate to
-// kh.cron_create/list/cancel which use the real kernel tick loop (#2024).
-
-async fn tool_schedule_create(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-    caller_agent_id: Option<&str>,
-    sender_id: Option<&str>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let agent_id = caller_agent_id.ok_or("Agent ID required for schedule_create")?;
-    let description = input["description"]
-        .as_str()
-        .ok_or("Missing 'description' parameter")?;
-    let schedule_str = input["schedule"]
-        .as_str()
-        .ok_or("Missing 'schedule' parameter")?;
-    let message = input["message"].as_str().unwrap_or(description);
-
-    let cron_expr = parse_schedule_to_cron(schedule_str)?;
-
-    // CronJob name only allows alphanumeric + space/hyphen/underscore (max 128 chars).
-    // Sanitize the user-provided description to fit these constraints.
-    let name: String = description
-        .chars()
-        .filter(|c| c.is_alphanumeric() || *c == ' ' || *c == '-' || *c == '_')
-        .take(128)
-        .collect();
-    let name = if name.is_empty() {
-        "scheduled-task".to_string()
-    } else {
-        name
-    };
-
-    // Build CronJob JSON compatible with kh.cron_create()
-    let tz = input["tz"].as_str();
-    let schedule = if let Some(tz_str) = tz {
-        serde_json::json!({ "kind": "cron", "expr": cron_expr, "tz": tz_str })
-    } else {
-        serde_json::json!({ "kind": "cron", "expr": cron_expr })
-    };
-    let mut job_json = serde_json::json!({
-        "name": name,
-        "schedule": schedule,
-        "action": { "kind": "agent_turn", "message": message },
-        "delivery": { "kind": "none" },
-    });
-    if let Some(obj) = job_json.as_object_mut() {
-        if !obj.contains_key("peer_id") {
-            if let Some(pid) = sender_id {
-                if !pid.is_empty() {
-                    obj.insert(
-                        "peer_id".to_string(),
-                        serde_json::Value::String(pid.to_string()),
-                    );
-                }
-            }
-        }
-    }
-
-    let result = kh
-        .cron_create(agent_id, job_json)
-        .await
-        .map_err(|e| e.to_string())?;
-    Ok(format!(
-        "Schedule created and will execute automatically.\n  Cron: {cron_expr}\n  Original: {schedule_str}\n  {result}"
-    ))
-}
-
-async fn tool_schedule_list(
-    kernel: Option<&Arc<dyn KernelHandle>>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let agent_id = caller_agent_id.ok_or("Agent ID required for schedule_list")?;
-    let jobs = kh.cron_list(agent_id).await.map_err(|e| e.to_string())?;
-
-    if jobs.is_empty() {
-        return Ok("No scheduled tasks.".to_string());
-    }
-
-    let mut output = format!("Scheduled tasks ({}):\n\n", jobs.len());
-    for j in &jobs {
-        let enabled = j["enabled"].as_bool().unwrap_or(true);
-        let status = if enabled { "active" } else { "paused" };
-        let schedule_display = j["schedule"]["expr"]
-            .as_str()
-            .or_else(|| j["schedule"]["every_secs"].as_u64().map(|_| "interval"))
-            .unwrap_or("?");
-        output.push_str(&format!(
-            "  [{status}] {} — {}\n    Schedule: {}\n    Next run: {}\n\n",
-            j["id"].as_str().unwrap_or("?"),
-            j["name"].as_str().unwrap_or("?"),
-            schedule_display,
-            j["next_run"].as_str().unwrap_or("pending"),
-        ));
-    }
-
-    Ok(output)
-}
-
-async fn tool_schedule_delete(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    // Accept either "id" or "job_id" for backward compatibility
-    let id = input["id"]
-        .as_str()
-        .or_else(|| input["job_id"].as_str())
-        .ok_or("Missing 'id' parameter")?;
-    kh.cron_cancel(id).await.map_err(|e| e.to_string())?;
-    Ok(format!("Schedule '{id}' deleted."))
-}
-
-// ---------------------------------------------------------------------------
-// Cron scheduling tools (delegated to kernel via KernelHandle trait)
-// ---------------------------------------------------------------------------
-
-async fn tool_cron_create(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-    caller_agent_id: Option<&str>,
-    sender_id: Option<&str>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let agent_id = caller_agent_id.ok_or("Agent ID required for cron_create")?;
-    let mut job = input.clone();
-    if let (Some(pid), Some(obj)) = (sender_id, job.as_object_mut()) {
-        if !pid.is_empty() && !obj.contains_key("peer_id") {
-            obj.insert(
-                "peer_id".to_string(),
-                serde_json::Value::String(pid.to_string()),
-            );
-        }
-    }
-    kh.cron_create(agent_id, job)
-        .await
-        .map_err(|e| e.to_string())
-}
-
-async fn tool_cron_list(
-    kernel: Option<&Arc<dyn KernelHandle>>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let agent_id = caller_agent_id.ok_or("Agent ID required for cron_list")?;
-    let jobs = kh.cron_list(agent_id).await.map_err(|e| e.to_string())?;
-    serde_json::to_string_pretty(&jobs).map_err(|e| format!("Failed to serialize cron jobs: {e}"))
-}
-
-async fn tool_cron_cancel(
-    input: &serde_json::Value,
-    kernel: Option<&Arc<dyn KernelHandle>>,
-    caller_agent_id: Option<&str>,
-) -> Result<String, String> {
-    let kh = require_kernel(kernel)?;
-    let job_id = input["job_id"]
-        .as_str()
-        .ok_or("Missing 'job_id' parameter")?;
-    let agent_id = caller_agent_id.ok_or("Agent ID required for cron_cancel")?;
-    // Authorize: the caller may only cancel jobs that belong to them.
-    // Otherwise an agent with the cron_cancel tool could delete any other
-    // agent's jobs as long as it learns their UUID (via side-channel or
-    // social engineering).
-    let owned = kh.cron_list(agent_id).await.map_err(|e| e.to_string())?;
-    let owns_job = owned.iter().any(|job| {
-        job.get("id")
-            .and_then(|v| v.as_str())
-            .is_some_and(|id| id == job_id)
-    });
-    if !owns_job {
-        return Err(format!(
-            "Cron job '{job_id}' not found or not owned by this agent"
-        ));
-    }
-    kh.cron_cancel(job_id).await.map_err(|e| e.to_string())?;
-    Ok(format!("Cron job '{job_id}' cancelled."))
 }
 
 // ---------------------------------------------------------------------------
